Redesigned list item model. Fixed overlapping label text in reused cells.

+++ MEADepthCamera/UI/Use Case/UseCaseListViewModel.swift
@@ -29,19 +29,21 @@ class UseCaseListViewModel: NSObject, ListViewModel {
         }
     }
     
+    weak var delegate: UseCaseInteractionDelegate?
+    
     var filter: UseCaseListViewModel.Filter = .all {
         didSet {
-            reloadListSectionStore()
+            reloadListStores()
         }
     }
     
     lazy var sectionsStore: ObservableModelStore<ListSection>? = {
-        guard let items = filteredUseCases?.compactMap({ ListItem(object: $0)?.id }) else { return nil }
+        guard let items = sortedUseCases?.compactMap({ $0.id }) else { return nil }
         return ObservableModelStore([ListSection(id: .list, items: items)])
     }()
     
     lazy var itemsStore: ObservableModelStore<ListItem>? = {
-        guard let items = filteredUseCases?.compactMap({ ListItem(object: $0) }) else { return nil }
+        guard let items = sortedUseCases?.compactMap({ listItem(useCase: $0) }) else { return nil }
         return ObservableModelStore(items)
     }()
     
@@ -56,22 +58,40 @@ class UseCaseListViewModel: NSObject, ListViewModel {
     }
 
     private var filteredUseCases: [UseCase]? {
-        return useCases?.filter { filter.shouldInclude(date: $0.date!) }.sorted { $0.date! > $1.date! }
+        return useCases?.filter { filter.shouldInclude(date: $0.date!) }//.sorted { $0.date! > $1.date! }
+    }
+    
+    private var sortedUseCases: [UseCase]? {
+        guard var useCases = filteredUseCases else { return nil }
+        let p = useCases.partition(by: { !Filter.today.shouldInclude(date: $0.date!) })
+        useCases[..<p].sort { $0.title! < $1.title! }
+        useCases[p...].sort { $0.title! < $1.title! }
+        return useCases
+    }
+    
+    func useCase(with id: UUID?) -> UseCase? {
+        return useCases?.first(where: { $0.id == id})
+    }
+    
+    func listItem(useCase: UseCase) -> ListItem? {
+        guard let id = useCase.id else { return nil }
+        let titleText = useCase.title ?? "?"
+        let subTitleText = useCase.experimentTitle ?? useCase.experiment?.title ?? "?"
+        let dateText = useCase.dateTimeText(for: filter) ?? "?"
+        let subjectID = useCase.subjectID ?? "?"
+        let subjectIDText = "Subject ID: " + subjectID
+        let bodyText = [subjectIDText, dateText]
+        return ListItem(id: id, title: titleText, subTitle: subTitleText, bodyText: bodyText)
     }
     
     func add(completion: @escaping (UseCase) -> Void) {
         dataProvider.add(in: dataProvider.persistentContainer.viewContext, shouldSave: false) { useCase in
-            self.addToStores(useCase)
             completion(useCase)
         }
     }
     
     func delete(_ useCase: UseCase, completion: @escaping (Bool) -> Void) {
-        guard let id = useCase.id else { return }
-        dataProvider.delete(useCase) { [weak self] success in
-            if success {
-                self?.deleteFromStores(id)
-            }
+        dataProvider.delete(useCase) { success in
             completion(success)
         }
     }
@@ -79,43 +99,68 @@ class UseCaseListViewModel: NSObject, ListViewModel {
 
 extension UseCaseListViewModel {
     
-    private func reloadListSectionStore() {
-        guard let itemIDs = filteredUseCases?.compactMap({ $0.id }) else { return }
-        sectionsStore?.merge(newModels: [ListSection(id: .list, items: itemIDs)])
+    private func sortUseCases() {
+        let todayFilter = Filter.today
+        
+        
     }
 
-    private func reloadItemStore() {
-        guard let items = filteredUseCases?.compactMap({ ListItem(object: $0) }) else { return }
-        reloadListSectionStore()
+    private func reloadListStores() {
+        guard let items = sortedUseCases?.compactMap({ listItem(useCase: $0) }),
+              let itemIDs = sortedUseCases?.compactMap({ $0.id })
+        else { return }
+        sectionsStore?.merge(newModels: [ListSection(id: .list, items: itemIDs)])
         itemsStore?.reload(with: items)
     }
     
-    private func addToStores(_ useCase: UseCase) {
-        guard let item = ListItem(object: useCase) else { return }
-        reloadListSectionStore()
+    private func addToListStores(_ useCase: UseCase) {
+        guard let id = useCase.id,
+              let item = listItem(useCase: useCase),
+              var listSection = sectionsStore?.fetchByID(.list)
+        else { return }
+        listSection.items?.append(id)
+        sectionsStore?.merge(newModels: [listSection])
         itemsStore?.merge(newModels: [item])
     }
     
-    private func deleteFromStores(_ itemID: UUID) {
-        reloadListSectionStore()
-        itemsStore?.deleteByID(itemID)
+    private func reconfigureItem(_ useCase: UseCase) {
+        guard let item = listItem(useCase: useCase) else { return }
+        itemsStore?.merge(newModels: [item])
     }
     
+    private func deleteFromListStores(_ itemID: UUID) {
+        guard var listSection = sectionsStore?.fetchByID(.list) else { return }
+        listSection.items?.removeAll(where: { $0 == itemID })
+        sectionsStore?.merge(newModels: [listSection])
+        itemsStore?.deleteByID(itemID)
+    }
 }
-
-
 
 // MARK: NSFetchedResultsControllerDelegate
 extension UseCaseListViewModel: NSFetchedResultsControllerDelegate {
+    
     func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
-//        print("controller did change object")
-//        fetchedResultsController(didChange: anObject, at: indexPath, for: type, newIndexPath: newIndexPath)
-//        print(#function)
-    }
-    func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-//        updateStores()
-//        applySnapshotFromListStore()
-//        print(#function)
+        guard let useCase = anObject as? UseCase,
+              let id = useCase.id
+        else { return }
+        switch type {
+        case .insert:
+            print("INSERT")
+            addToListStores(useCase)
+        case .delete:
+            print("DELETE")
+            deleteFromListStores(id)
+        case .move:
+            print("MOVE")
+            reloadListStores()
+            delegate?.didUpdateUseCase(useCase)
+        case .update:
+            print("UPDATE")
+            reconfigureItem(useCase)
+            delegate?.didUpdateUseCase(useCase)
+        @unknown default:
+            reloadListStores()
+        }
     }
 }
 
@@ -136,203 +181,8 @@ extension UseCaseListViewModel: UISearchResultsUpdating {
         } catch {
             fatalError("###\(#function): Failed to performFetch: \(error)")
         }
-
-        reloadItemStore()
+        
+        reloadListStores()
     }
 }
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-class OldUseCaseListViewModel: NSObject, OldListViewModel {
-
-    typealias ListDiffableDataSource = UICollectionViewDiffableDataSource<ListSection.ID, ListItem.ID>
-    typealias ListCell = ListTextCell
-    typealias HeaderCell = UseCaseListCell
-    
-    enum Filter: Int {
-        case today
-        case past
-        case all
-
-        func shouldInclude(date: Date) -> Bool {
-            let isInToday = Locale.current.calendar.isDateInToday(date)
-            switch self {
-            case .today:
-                return isInToday
-            case .past:
-                return (date < Date()) && !isInToday
-            case .all:
-                return true
-            }
-        }
-    }
-    
-    var dataSource: ListDiffableDataSource?
-    
-    var filter: OldUseCaseListViewModel.Filter = .all
-    
-    lazy var sectionsStore: ObservableModelStore<ListSection>? = {
-        guard let items = filteredUseCases?.compactMap({ ListItem(object: $0)?.id }) else { return nil }
-        return ObservableModelStore([ListSection(id: .list, items: items)])
-    }()
-    
-    lazy var itemsStore: ObservableModelStore<ListItem>? = {
-        guard let items = filteredUseCases?.compactMap({ ListItem(object: $0) }) else { return nil }
-        return ObservableModelStore(items)
-    }()
-    
-    private lazy var dataProvider: UseCaseProvider = {
-        let container = AppDelegate.shared.coreDataStack.persistentContainer
-        let provider = UseCaseProvider(with: container, fetchedResultsControllerDelegate: self)
-        return provider
-    }()
-    
-    private var useCases: [UseCase]? {
-        return dataProvider.fetchedResultsController.fetchedObjects
-    }
-
-    private var filteredUseCases: [UseCase]? {
-        return useCases?.filter { filter.shouldInclude(date: $0.date!) }.sorted { $0.date! > $1.date! }
-    }
-    
-    private lazy var allItemsSubscriber: AnyCancellable? = {
-        return itemsStore?.$allModels
-            .receive(on: RunLoop.main)
-            .sink { [weak self] _ in
-                // TODO: apply changes to snapshot
-            }
-    }()
-    
-    func configure(_ listCell: ListTextCell) {
-        listCell.delegate = self
-    }
-    
-    func update(_ useCase: UseCase, completion: (Bool) -> Void) {
-        if let context = useCase.managedObjectContext {
-            dataProvider.persistentContainer.saveContext(backgroundContext: context, with: .updateUseCase)
-//            context.refresh(useCase, mergeChanges: true)
-            completion(true)
-        } else {
-            completion(false)
-        }
-    }
-    
-    func add(completion: @escaping (UseCase) -> Void) {
-        dataProvider.add(in: dataProvider.persistentContainer.viewContext, shouldSave: false) { useCase in
-            completion(useCase)
-        }
-    }
-}
-
-extension OldUseCaseListViewModel {
-    
-    private func reloadSectionStore() {
-        guard let itemIDs = filteredUseCases?.compactMap({ $0.id }) else { return }
-        sectionsStore?.reload(with: [ListSection(id: .list, items: itemIDs)])
-    }
-
-    private func reloadItemStore() {
-        guard let items = filteredUseCases?.compactMap({ ListItem(object: $0) }) else { return }
-        itemsStore?.reload(with: items)
-    }
-
-    private func mergeIntoItemStore(_ useCase: UseCase) {
-        guard let item = ListItem(object: useCase) else { return }
-        itemsStore?.merge(newModels: [item])
-    }
-    
-    private func deleteFromItemStore(_ useCase: UseCase) {
-        guard let itemID = useCase.id else { return }
-        itemsStore?.deleteByID(itemID)
-    }
-    
-}
-
-// MARK: ListTextCellDelegate
-extension OldUseCaseListViewModel: ListTextCellDelegate {
-    
-    func contentConfiguration(for item: ListItem) -> TextCellContentConfiguration? {
-        guard let useCase = item.object as? UseCase else { fatalError() }
-        guard let titleText = useCase.title,
-              let experimentText = useCase.experimentTitle,
-              let dateText = useCase.dateTimeText(for: .all),
-              let subjectID = useCase.subjectID
-        else { return nil }
-        let subjectIDText = "Subject ID: " + subjectID
-        let completedTasksText = "X out of X tasks completed"
-        let bodyText = [subjectIDText, dateText, completedTasksText]
-        let content = TextCellContentConfiguration(titleText: titleText, subtitleText: experimentText, bodyText: bodyText)
-        return content
-    }
-    
-    func delete(objectFor item: ListItem) {
-        guard let useCase = item.object as? UseCase else { fatalError() }
-        dataProvider.delete(useCase) { [weak self] success in
-            if let dataSource = self?.dataSource {
-//                var snapshot = dataSource.snapshot()
-//                snapshot.deleteItems([item.id])
-//                dataSource.apply(snapshot)
-                // TODO: delete from item store
-            }
-        }
-    }
-}
-
-// MARK: UseCaseInteractionDelegate
-extension OldUseCaseListViewModel: UseCaseInteractionDelegate {
-    /**
-     didUpdateUseCase is called as part of UseCaseInteractionDelegate, or whenever a use case update requires a UI update (including main-detail selections).
-     
-     Respond by updating the UI as follows.
-     - add:
-     - delete: reload snapshot and apply to collection view data source
-     - update from detailViewController:
-     - initial load:
-     */
-    func didUpdateUseCase(_ useCase: UseCase) {
-        guard let itemID = useCase.id else { fatalError() }
-        if let dataSource = dataSource {
-//            var snapshot = dataSource.snapshot()
-//            snapshot.reconfigureItems([itemID])
-//            dataSource.apply(snapshot)
-        }
-    }
-}
-
-// MARK: UISearchResultsUpdating
-extension OldUseCaseListViewModel: UISearchResultsUpdating {
-    func updateSearchResults(for searchController: UISearchController) {
-    }
-}
-
-// MARK: NSFetchedResultsControllerDelegate
-extension OldUseCaseListViewModel: NSFetchedResultsControllerDelegate {
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
-//        print("controller did change object")
-        fetchedResultsController(didChange: anObject, at: indexPath, for: type, newIndexPath: newIndexPath)
-//        print(#function)
-    }
-    func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-//        updateStores()
-//        applySnapshotFromListStore()
-//        print(#function)
-    }
-}
